blockchain: Unsigned treasury spend vote counts.
As mentioned in the previous commit, consensus code should ideally
always use types of a specific size so there is no possibility of
divergent behavior when compiled on different architectures due to
differing upper bounds.  Further, unsigned types for values that can
never be negative should always be preferred.

The current code for counting treasury spend votes does not adhere to
practice and uses plain signed integers.

It is worth noting that this is not an active issue at the moment
because the maximum possible counts achievable due to other limits
imposed on the max number of votes and the window over which the number
of votes are counted are less than the maximum value of the smallest
supported architecture (32 bits).

Nevertheless, these types of hidden assumptions are fragile and can
easily lead to undetected and unexpected behavior when parameters
change and new features are introduced.

The primary goal of this commit is to address that by modifying the
relevant code to use fixed unsigned 32-bit integers instead.

While here, it also takes the opportunity to cleanup the code to make it
more consistent with the other consensus code.

Note that there is no risk of consensus divergence due to the
aforementioned impossibility of hitting the conditions with the current
combination of parameters and limits.

+++ internal/rpcserver/interface.go
@@ -440,7 +440,7 @@ type Chain interface {
 
 	// TSpendCountVotes returns the votes for the specified tspend up to
 	// the specified block.
-	TSpendCountVotes(*chainhash.Hash, *dcrutil.Tx) (int64, int64, error)
+	TSpendCountVotes(*chainhash.Hash, *dcrutil.Tx) (uint32, uint32, error)
 
 	// InvalidateBlock manually invalidates the provided block as if the block
 	// had violated a consensus rule and marks all of its descendants as having
